2353-fix Filters top padding, scrollbar, and 1px gap
